services/finance-langgraph/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
pydantic==2.9.2
langgraph==0.2.39
langchain-core==0.3.9
google-cloud-bigquery==3.25.0
